exfatprogs: Fix integer signedness and width confusions
Integer type confusions are foudn with the following build
configurations:

    CC="clang" CFLAGS="-m32 -Wall -Wextra" ./configure
    CC="clang" CFLAGS="-m64 -Wall -Wextra" ./configure
    CC="gcc"   CFLAGS="-m32 -Wall -Wextra" ./configure
    CC="gcc"   CFLAGS="-m64 -Wall -Wextra" ./configure

Handle rare cases of integer overflows where needed. Fix benign type
confusions(size_t vs int).

+++ mkfs/upcase.c
@@ -30,7 +30,7 @@ int exfat_create_upcase_table(struct exfat_blk_dev *bd,
 	 * TODO do this in a loop.
 	 */
 	nbytes = pwrite(bd->dev_fd, ui->upcase.table, ui->upcase.len, finfo.ut_byte_off);
-	if (nbytes != ui->upcase.len)
+	if (nbytes != (ssize_t)ui->upcase.len)
 		return -1;
 	if (ui->verify) {
 		ret = exfat_check_written_data(bd,
